requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
numpy
opencv-python-headless
fastapi
uvicorn
python-multipart
pydantic
pillow
requests
